refactor: remove ImpersonationAlert component and related toast logic; introduce Wrapper component for toast handling
