+++ resources/js/components/ui/sonner/index.ts
@@ -1,1 +1,1 @@
-export { default as Toaster } from './Sonner.vue';
+export { default as Toaster } from './Wrapper.vue';
